bpf: Remove hash function
Instead of using the hash function, use strings and if necessary, wrap them
in structures.
